refactor: move dependency folder to locator
