--- lib/src/di/register_manager_module.dart
@@ -1,8 +1,0 @@
-part of 'di_graph_setup.dart';
-
-void _registerManagersModule() {
-  // Get.put(ConnectivityManager());
-  // Get.put(PermissionManager());
-  // Get.put(TranslationManager(Get.find()));
-  // Get.put(DeepLinkManager());
-}
