+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
     sdk: flutter
   flutter_bloc: ^8.1.1
   flutter_lorem: ^2.0.0
+  flutter_native_splash: ^2.2.16
   flutter_slidable: ^2.0.0
   flutter_vector_icons: ^2.0.0
   fluttertoast: ^8.1.1
@@ -58,8 +59,8 @@ dependencies:
 dev_dependencies:
   built_value_generator: ^8.4.1
   flutter_launcher_icons: ^0.11.0
-  flutter_native_splash: ^2.2.16
   flutter_lints: ^2.0.0
+  flutter_native_splash: ^2.2.16
   flutter_test:
     sdk: flutter
   hive_generator: ^2.0.0
@@ -70,8 +71,6 @@ flutter_native_splash:
   image: assets/images/logo.png
   color_dark: "#ffffff"
   image_dark: assets/images/logo.png
-  
-
 
 flutter_icons:
   android: "launcher_icon"
